requirements.txt
# Gmail OAuth Integration Requirements
# Python 3.7+

# Qt GUI Framework
PyQt6>=6.4.0

# Google API Libraries
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Database Support
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# HTML Sanitization (Optional, but recommended)
html-sanitizer>=2.0.0

# HTTP Client for Gateway Integration
requests>=2.28.0

# Utilities
python-dateutil>=2.8.0

# Windows Desktop Icon Support (Optional)
pywin32>=305; sys_platform == "win32"

# Image Processing for Reverse Image Search
Pillow>=9.0.0

# Reverse Image Search (Optional - for phishing detection)
selenium>=4.0.0
webdriver-manager>=3.8.0

